Replace corr2() function
The standard function to calculate a 2D correlation coefficient does a
lot of input checking.  As it is called in a loop, a lot of redundant
checks are done.  As we know the input images are compatible, it is
faster to implement this simple function ourselves.

Also add unittest to check whether implementation is identical to the
corr2() function.  As calculate_correlation_map() is a local function,
some trickery in the main piv_FFTmulti() function is needed to expose
it to the test suite.  At some point this needs to be packaged correctly
in a separate .m file.