Don't store full X1/Y2 matrices
These are only input to U1/V1, so they don't need to take up space after that.
The X1/Y1 matrices are still there, but only 1D vectors now, so not much memory is wasted.

This is also a performance improvement, as interp2 works faster on two
1D vectors for X1/Y2 instead of the repmat-expanded 2D matrices.

The interp2() before the repmat() call is an optimization for temporary
allocations.  When the repmat() is the left operand of `+` and interp2()
the right operand, there will be an additional allocation.