Store mask as logical
It is only used in a find call, so it only matters whether the values
are nonzero or not.  Furthermore, the poly2mask function already returns
a logical array, so it can easily be or-ed with the mask if it is also logical.

In a logical array, each element is 1 byte (not 1 bit), so 7/8 of `mask`
and `gen_mask` is saved compared to storing it as double.