Always enable calculation of the correlation map
This is not that expensive anymore and in some later commit we will not have
access to the full image_cut arrays at the end of the computation anymore,
so the call to calculate_correlation_map needs to be done early on.

The calculated correlation map might change a bit in some cases.
The correlation map is now always done on the main image_cut, not on the
last one of the repeated multishift correlation.  Also, the correlation
map is now calculated before doing any padding on the smaller images.
Both changes make sense and are arguably more correct.