Store large arrays with indices as uint32
This saves memory and processing time.

For now, using uint32 for indexing means limiting images to be at most
in the gigapixel range.  In a later commit though, linear indexing will
not be used anymore and the uint32 limit applies only to the image size
in a single direction.

Additionally, the typevector can be a uint8 array instead of double.