Apply mask with simple indexing
The ss1 array is not needed for masking when simply indexing mask with X and Y coordinates.
This will be useful later when the ss1 array is eliminated entirely.

The strange indexing of result_conv with mask(masked_ys-1, masked_xs+1)'
is to preserve old results.  But that should probably be changed to
simply mask(masked_ys, masked_xs).