Only do padding of images once
It does not make sense to go back to the original image and pad it again
every pass, each time with a smaller interrogation area.
Instead, just pad it once to the first interrogation area size, which is
the largest.  Then in the passes, the padded images are indexed such
that the original behavior is kept.