Always work with x/y in input image coordinates
Instead of adjusting xtable/ytable at the end with padding and roi
offset, use the same reference throughout the piv_FFTmulti function.
This makes it clearer what is happening by removing some padding offset
shifts.

This reveils a possible improvement in the interpolation of
xtable/ytable from one pass to the next.