Transpose 3rd dimension of result_conv
The 3rd dimension of result_conv previously had the all the subimages with
the x-direction increasing fastest.  This is now fastest in the y-direction.
The 3rd dimension of result_conv now corresponds to linear indexing into
the typevector and correlation_map arrays, which is more consistent.

The results change a bit due to different rounding in the FFT stages,
but only where there is not a clear correlation anyway.