Don't convert image to double first
We can immediately convert from the input datatype to the image class
type we want to do the processing on.

As a result, the images are now always processed in floating point with
pixel values between 0 and 1.  Previously they were processed in
floating point between 0 and the maximum integer value of the input
datatype (e.g. 255 for uin8).
This should affect results only in a roundoff error.