Improve performance of calculate_correlation_map()
Having the testcase for this function gives a bit more confidence in the
correctness of the performant, but slightly difficult to read implementation.

The main reason for the improved performance compared to the previous
implementation is that the calculation of the means is done for all
images in one statement and that we avoid the copy that happened on the
`a = img1(:,:,i)` lines.